Add username in JWT returned by keycloak
In production SSO, the JWT contains a username field that contains the email of the user.
We use this field to create the user if he does not exist.

This patch change the username from dci to dci@redhat.com and add the username field in the JWT returned by keycloak.

Change-Id: Iffe1f0890bd11932611edc76b489ff38a32a3717